+++ tools/level_editor/level_editor.vala
@@ -1348,6 +1348,7 @@ public class LevelEditorApplication : Gtk.Application
 		this.set_accels_for_action("database.undo", new string[] { "<Primary>Z" });
 		this.set_accels_for_action("database.redo", new string[] { "<Shift><Primary>Z" });
 		this.set_accels_for_action("database.duplicate", new string[] { "<Primary>D" });
+		this.set_accels_for_action("project-browser.delete", new string[] { "Delete" });
 		this.set_accels_for_action("database.delete", new string[] { "Delete" });
 		this.set_accels_for_action("database.select_none", new string[] { "<Shift><Primary>A" });