+++ docs/changelog.rst
@@ -22,6 +22,7 @@ Changelog
 * Tools: added the ability to filter projects in the Projects List.
 * Tools: resource inputs for units, textures and materials now show a thumbnail and a tooltip with additional info.
 * Tools: add 'New Package...' menu entry in the Project Browser.
+* Tools: added the ability do delete resources in the Project Browser via ``Delete`` keyboard shortcut.
 * Runtime: added support for MP3 sound files.
 * Runtime: added triplanar mapping support to the mesh shader (+TRIPLANAR and +TRIPLANAR_LOCAL).
 * Runtime: added a new Vignette post-processing effect.